+++ swervelib/src/main/java/frc/team1711/swerve/commands/AutonDriveSimple.java
@@ -1,7 +1,3 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
 package frc.team1711.swerve.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
