+++ swervelib/src/main/java/frc/team1711/swerve/subsystems/AutoSwerveWheel.java
@@ -5,9 +5,9 @@
 package frc.team1711.swerve.subsystems;
 
 /**
- * Used by {@link AutoSwerveDrive} to represent a module with encoders on the drive motors,
- * in addition to the encoders on the steering motors which are necessary for basic swerve
- * functionality.
+ * Used by {@link AutoSwerveDrive} to represent a module with encoders on the drive motors
+ * (in addition to the encoders on the steering motors which are necessary for basic swerve
+ * functionality).
  * @author Gabriel Seaver
  */
 abstract public class AutoSwerveWheel extends SwerveWheel {
